[iOS][QD] Calculate tabs to be closed before dismissing the bottomsheet
We noticed that on real devices, there could be some noticeable delay
between dismissing Quick Delete and the animation starting. We believe
this was because it can take some time a bit to calculate the tabs to
close. In this CL, we calculate these tabs before dismissing the
animation so the delay to trigger the animation is not noticeable.

Fixed: 358568193
Bug: 335387869, 354112735
Change-Id: I7ace4b5c8876bdf845bc165f10691e9c105d09fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5776949
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Filipa Senra <fsenra@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1341613}