Remove FormEventLoggerBase::signin_state_for_metrics_
The base class shouldn't contain a field of type PaymentsSigninState,
but instead live in the CreditCardFormEventLogger.

This CL also removes PaymentsSigninState parameters from various
FormEventLoggerBase functions. Since all these functions simply set the
member as their first operation, this is moved to the caller side; in
the BAM, whenever the credit card logger gets called, it now calls a
setter for the metrics state first.

One write reference in AddressFormEventLogger for the field exists too.
But this reference is irrelevant, since the address form logger is only
called for address forms. Because the field is only read by the card
event logger (which is only triggered for payments forms), the write
is never read.

Bug: 40100455
Change-Id: If7188b71ab5061551adb49914ef56d63a0d1646e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5785821
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1341630}