DevTools: Add Extensions.setStorageItems CDP command
This CL adds an Extensions.setStorageItems CDP command which can be
used to set values in the storage areas associated with a Chrome
extension.

The original proposal document had a non-plural API that could only set
a single key, but that seemed like an unnecessary restriction, so having
this match the general signature of the extension API where multiple
keys can be set in one command.

Bug: 40963428
Change-Id: I2c742462da90b2d5218ec46fb60877cdad53efbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5783237
Commit-Queue: Oliver Dunk <oliverdunk@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341636}