Change Observer: Fix crash when navigating to new page
The browser may crash if you navigate to a new page with an active
FileSystemObserver observation.

This is because the FileSystemAccessObserverObservation is a
FileSystemAccessPermissionGrant::Observer of the root handle's read
grant. And when the permission status changes, it destroys itself along
with the read grant if its the only holder of the read grant.

This means that when you navigate to a new page, RevokeActiveGrants
attempts to iterate through all the active grants of the page and revoke
them. But since its iterating over an array that
FileSystemAccessObserverObservation removes an element from during that
iteration, it get an invalid iterator.

The fix is to increase the iterator before continuing.

Fixed: 359567614
Change-Id: I3eea7340527e0255bdaa7488cf6e4968b299483e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5786874
Reviewed-by: Daseul Lee <dslee@chromium.org>
Commit-Queue: Nathan Memmott <memmott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341696}