[AvatarButton] Fix flaky browser tests that are time dependent
Previously tests for the AvatarToolbarButton were relying on changing
the duration of delays enough to be able to test the behavior but at
the same time not too much in order not to create long waiting tests.
This caused consistent flakiness in tests, especially that the
different OS uses different machines (extreme flakiness on Windows for
example). Even though events notifications were used, it was still
not very accurate, with sometimes delays finishing too early for
example.

This CL introduces a different logic, which gives 3 different testing
functions for each delay:
- Increase the delay for a long duration that guarantees that testing
the behavior is safe and sure.
- Disable a delay, this way you can bypass unnecessary waiting time
without any behavior change.
- Stopping a delay (enforcing the ending callback and stopping the
timer/task), to be used when transitioning from a waiting mode, to the
post task behavior that is expected.

This approach may make the test less accurate as it will test a
slightly different behavior than the real once (not actually waiting
at all anymore, but instead triggering it ourselves). However, this is
acceptable given that we mainly want to test the pre/post behavior of
a delay.

Re-enabled all tests that were disabled due to flakiness (some were
only disabled on Windows).

Removed a test that was only time dependent where the user signs in
and out multiple times. The test was testing some edge cases that were
time dependent which should be fine not to have as a test after all.

Fixed: 327688158, 330202396
Fixed: b:331746545
Change-Id: I17a214c6ee6d9d489592a450e688bd2d934fd658
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5741427
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1341722}