[views-ax] Migrating AboutThisSiteSidePanel view to new system
This CL migrates the AboutThisSiteSidePanel view to use the new system
for accessible attributes.

We use the name and role that are being exposed by `WebView` as those
are the ones we were exposing before.

We can't leave as is, because we'll need to migrate the call
to `web_view_->GetAccessibleNodeData` to use
`web_view_->GetViewAccessibility().GetAccessibleNodeData()` since
that is the correct pattern, but in this case we would then end up
with ALL the a11y attributes from the `WebView` being exposed in
`AboutThisSiteSidePanel` as well, which is undesirable.

Bug: 325137417

Change-Id: I23e00210ddaa6ec0761c785cf0d71b1af0bfab1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5783976
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1341724}