[lensoverlay] Continue monitoring underneath web contents in background
Previously, we stopped observing the underlying web contents if we
entered the background. However, this meant if something happened to
the underlying web contents, while we were in the previous tab, the
overlay would stay open in a bad state. Now, in the background, we keep
the listener to:
  1) Detect renderer crashes
  2) Detect URL changes

Bug: b:357977717, b:357648055, b:359678194
Change-Id: Iee8f9ab2532c4c1107512af11cfd8777dd1970f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5787337
Reviewed-by: Juan Mojica <juanmojica@google.com>
Commit-Queue: Duncan Mercer <mercerd@google.com>
Cr-Commit-Position: refs/heads/main@{#1341729}