[iOS TG Sync] Implement IsSaveableNavigation for iOS
This CL adds various conditions to the navigation to determine
whether a navigation should be synced. This is intended to be a
mirror of the existing TabGroupSyncUtils::IsSaveableNavigation on
desktop and android.

The next CL https://chromium-review.googlesource.com/c/chromium/src/+/5788102
 filters out unsupported schemes from syncing and replaces them
with an NTP wherever required.

This also required slight change in signature of
FakeNavigationContext::SetWebState to accept a pointer instead
of std::unique_ptr.

Bug: 359726089
Change-Id: Ic41cb3f4a8c5a6d24a4e99cf4b1f955e59e54f0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5787832
Reviewed-by: Min Qin <qinmin@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341761}