[ios] Change Tips Notifications trigger logic using classification
Previously, whenever the app came into the foreground the notification
client would check if there was a pending request and if there was, it
would remove it and request a new notification, thus extending the time
until one would trigger. If the user continued to come back into the
app the notifications would never trigger.

This new logic will not remove existing requests. It simply checks if
there is a pending request and if there isn't, then it will set up a
new request with a trigger time that depends on the user's
classification.

Fixed: 353984939
Change-Id: I0fad686678490e5fea5078d6eaa72985e587a602
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5781787
Commit-Queue: Scott Yoder <scottyoder@google.com>
Reviewed-by: Hira Mahmood <hiramahmood@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1341776}