Allow persistent passthrough of USB devices to guest OSes
At the moment, when users of guest OSes of chromeOS would like to pass
through a USB peripheral, they must manually do so each time the
peripheral is enumerated - be that via plugging it in, peripheral
rebooting, etc. At best, this can be annoying to have to do each time,
and at worst it makes certain use cases impossible - e.g. flashing a DUT
using servo.

This patch introduces a toggle that enables persistent passthrough, and
also the accompanying behaviors:
* If persistence is enabled, connecting a device will add it to a
  pref-based mapping of device->guest.
* Device identification is done on a best-effort basis, using the
  device's VID/PID, and a serial number if it is available.
* If a device is manually disconnected via the settings UI (as opposed
  to simply unplugged), that device is then also removed from the
  persistence mapping.
* If persistence is disabled, it also clears the existing mapping.

The persistence setting is available on subpages for guests, even though
the setting is a global one - there is no suitable global setting page
or subpage for guest OS behavior at this time, and resources to create
one are unlikely to materialize.

Screenshots:
* Top level toggle
  https://screenshot.googleplex.com/5cW8UfCWsTdB5aR
* Enabling
  https://screenshot.googleplex.com/58YDVGboYs5Gt7C
* Disabling
  https://screenshot.googleplex.com/4LYy5BtKDajZ5WT

Change-Id: I1477333a82a002a48c0d18e3faeb394e4c517236
Bug: 354016718
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5745757
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: maciek swiech <drmasquatch@google.com>
Reviewed-by: Wes Okuhara <wesokuhara@google.com>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Cr-Commit-Position: refs/heads/main@{#1341787}