[iOS] Handle TabGrid Done button through commands
This CL moves the handling of the Done button of the TabGrid out of the
TabGridToolbarsMainTabGridDelegate protocol and move it to the TabGrid
Commands protocol.

It allows to remove the plumbing of the delegate through the grids.

Bug: None
Change-Id: I5ecbb10e7c6e2830df1d17492fa75104f734adf5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5783210
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Aliona Dangla <alionadangla@chromium.org>
Reviewed-by: Louis Romero <lpromero@google.com>
Cr-Commit-Position: refs/heads/main@{#1341825}