[views-ax] Introduce lazy loading mechanism for Views accessibility
This CL introduces the function View::OnAccessibilityInitializing.
This function gives Views authors the power to decide what accessibility attributes should be loaded only when accessibility is
required, giving them the tools to make the right decision when it comes to performance and memory.

Accessibility attributes that involve complex computation should
leverage this lazy loading mechanism to avoid doing expensive work for
users who do not need the accessibility attributes. The same goes for
attributes that occupy a lot of memory space.

The mechanism starts with
ViewAccessibility::CompleteCacheInitialization, a function called from
View::AddChildViewImpl (or Widget::Init for the RootView only), and in
turns call View::OnAccessibilityInitializing.

We added two types of guardrails (checks) to help Views authors use
this mechanism properly:
1. There's a check that prevents the author from using a
   ViewAccessibility cache setter from the OnAccessibilityInitializing
   override.
2. There are DCHECKS that validate the lazy loaded attribute hasn't
   been loaded before, which would defeat its purpose.

For now, we always fully initialize the accessibility cache. In a
follow-up CL, we will limit it to only doing when the kNativeAPIs mode
is set.

Bug: 325137417
Change-Id: Id614bf9549a2d11cc0a8aa99b20d8a8a0c714fbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5668503
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: Javier Contreras <javiercon@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1341827}