Add metric for Private Aggregation budgeting contributions individually
We are considering performing budgeting on a per-contribution basis
rather than a per-report basis (i.e. if there's not enough budget for
the entire report, we could still allow certain contributions to be
sent). This metric will help understand how useful this would be.

Bug: 355271550
Change-Id: I0fd5c9eeacacd0ded22d3bc6856c1defa7575768
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5777485
Reviewed-by: John Delaney <johnidel@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Reviewed-by: Dan McArdle <dmcardle@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341873}