Revise dropping multiple URLs on the tab strip
The original CL that allowed a multi-item drop on the tab strip,
https://crrev.com/c/5350662, restricted the drop to fall between tabs
only. This was a bit confusing, and broke the "hover to switch tabs"
feature of the tab strip.

Therefore, allow dropping multiple items on tabs, and replace the
dropped-upon tab with multiple tabs when requested.

Bug: 41403933
Change-Id: I6177faaf459d0c22c124512731c69aace75f5be1
Fixed: 344943217
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5787493
Commit-Queue: Avi Drissman <avi@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: David Pennington <dpenning@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341874}