Support OpenFile current_folder in SelectFileDialogLinuxPortal
This change adds support for opening a file dialog in the current folder
in the SelectFileDialogLinuxPortal. Support for this feature was added
in xdg-desktop-portal via https://github.com/flatpak/xdg-desktop-portal/pull/1045.

Electron makes use of this feature and this allows us to remove a patch to enable this functionality.

Change-Id: I3f95bdff0c6a834f965ed3c153eb1a99902f61bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5786982
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341908}