[Test Automation] Fix default 'must remain visible' before subsequence
When not specified, for a step that references an element, there is
complex logic that determines whether the element should remain visible
through the triggering of the following step. For example, if the
following step requires the same element to be hidden, the default is
false. But if the following step references the same element and then
does something with it, the default is true.

The problem arises when a kShow step is followed by a kSubsequence
step. Because subsequences don't officially "trigger" until all of the
required subsequences complete, if in a (potentially long and complex)
subsequence the element is hidden, the previous step fails. This is
not expected behavior; once the subsequences start executing there
should be no expectation (at least by default) that the previous step's
element remains visible.

This CL updates the calculation of the default so that in this case, if
not otherwise specified, the flag will now default to false in this
case.

Change-Id: Ia28ef7e11af91545716db92d291344697d871a33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5778342
Commit-Queue: Dana Fried <dfried@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Joseph Park <josephjoopark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341914}