Hide bottom chin when showing the keyboard
Other bottom browser controls hide when showing the chin, and the bottom chin not currently doing this means that the composited layer shows above the keyboard accessory, at the bottom of the content. This change fixes that.

Fix:
https://screencast.googleplex.com/cast/NTAwMjU3OTM3NjM0MDk5MnwyNTNmMTJlZC00MA

Bug:
https://screenshot.googleplex.com/3wkqEuYiXdP6fa6.png
https://screencast.googleplex.com/cast/NjYwMDI3MDIzMzQ2ODkyOHw3NTc4OTE2Mi0xMw

Bug: 359669361
Change-Id: Iee45d4a40ce71ee45c39cc82bfb11c68d4a3fc17
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5787466
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Charles Hager <clhager@google.com>
Cr-Commit-Position: refs/heads/main@{#1341956}