[iOS] Toggle customization pref from settings for feed
A new pref was added to track feed visibility from the customization
menu, but this would force the feed to be enabled for all users opted
into the experiment, even if they had chosen to disable the feed
previously.

In order to avoid this, this CL uses the existing pref in the
customization menu. The "articles visible" pref refers to collapsing
the feed from the header menu which is no longer supported, so it is
ignored when the flag is enabled.

Bug: 350990359
Change-Id: Ib8ea3598d48d091fd4d4a5b51df8bdbd7cce31dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5784042
Commit-Queue: Adam Arcaro <adamta@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Auto-Submit: Adam Arcaro <adamta@google.com>
Cr-Commit-Position: refs/heads/main@{#1341976}