[Privacy Guide] Fix animation when navigating between cards
When cleaning up the PG-pre-v2 codepath (crrev.com/c/4120509) it was
overlooked that some animations were gated on the privacy-guide-page
element having a "v2" attribute. This attribute was removed with the
cleanup. As a result, the animations when navigating between cards
broke: they showed the new card content too early, which meant that
for a short time the content of both the old and new card was visible.

This CL fixes this by removing the "v2" gate for those animations.

Fixed: 355152082
Bug: 40184479
Change-Id: Ie7148bdca25df76a7061cea9feee57dfeaac8106
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5783240
Reviewed-by: John Lee <johntlee@chromium.org>
Commit-Queue: John Lee <johntlee@chromium.org>
Auto-Submit: Rainhard Findling <rainhard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341982}