[Extensions] Resolve pylint errors in schema compiler PRESUBMIT.py file
This CL adds comments to allow the toplevel import used in the PRESUBMIT
file and removes itself from the list of files to ignore for the pylint
checks.

Also updates the PRESUBMIT file to use the 2.0.0 version format, which
lets us remove some of the redundancy in the file. With this format all
functions beginning with "Check" are run automatically and do not need
to be called inside CheckChangeOnUpload/Commit.

Bug: 359623676
Change-Id: I981fc4f4d7c0798c51244776468a5a684418b090
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5784718
Commit-Queue: Tim <tjudkins@chromium.org>
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1342010}