[PWAs] Unbreak creating shortcuts for default apps on macOs.
In http://crrev.com/c/5640638 how shortcuts are created for default
apps was changed. However while the win/linux version of
ShowCreateChromeAppShortcutsDialog was updated in that CL, the mac
version wasn't, causing this functionality to break. While this was
caught by tests, the test was disabled and ignored instead of
investigated...

Bug: 349628411
Change-Id: I2022e35637212c9476642247ef23a0a44e3ebdaa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5789440
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1342026}