[Layout] Fix ToolbarController visibility oscillation
This changes how LayoutManagerBase tracks inclusion in the layout, which
in turn allows for temporary removal of a child view from the layout (in
this case, to determine if the toolbar should show an overflow
indicator).

Instead of setting the overflow button to hidden and then performing
the first pass of the layout, it now excludes it from the layout, and
only afterwards updates the visibility.

This in turn debounces the visibility state, making the overflow
button safe to access in e.g. IPH and Kombucha tests.

This CL also removes the redundant "decorative" property on some
Views, replacing it with "ignored by layout", as the two did (or at
least *should have done*) the same thing.

Bug: 41495158, 40939901
Change-Id: I7d2fa88457569e69c977175e89a6c8efb7b3cf35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5787103
Reviewed-by: pengchao Cai <pengchaocai@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: pengchao Cai <pengchaocai@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1342066}